+++ src/rgw/CMakeLists.txt
@@ -22,8 +22,6 @@ find_package(ICU 52.0 COMPONENTS uc REQUIRED)
 
 set(librgw_common_srcs
   services/svc_finisher.cc
-  services/svc_bi_rados.cc
-  services/svc_bilog_rados.cc
   services/svc_bucket.cc
   services/svc_bucket_sobj.cc
   services/svc_bucket_sync_sobj.cc
@@ -45,6 +43,8 @@ set(librgw_common_srcs
   services/svc_user.cc
   services/svc_user_rados.cc
   services/svc_zone.cc
+  services/rados/svc_bi_rados.cc
+  services/rados/svc_bilog_rados.cc
   services/rados/svc_rados.cc
   services/svc_zone_utils.cc
   rgw_service.cc
