+++ src/rgw/rgw_admin.cc
@@ -61,7 +61,7 @@ extern "C" {
 
 #include "services/svc_sync_modules.h"
 #include "services/svc_cls.h"
-#include "services/svc_bilog_rados.h"
+#include "services/rados/svc_bilog_rados.h"
 #include "services/svc_mdlog.h"
 #include "services/svc_meta_be_otp.h"
 #include "services/svc_zone.h"
