+++ src/rgw/rgw_bucket.cc
@@ -36,7 +36,7 @@
 #include "services/svc_meta_be_sobj.h"
 #include "services/svc_user.h"
 #include "services/svc_cls.h"
-#include "services/svc_bilog_rados.h"
+#include "services/rados/svc_bilog_rados.h"
 
 #include "include/rados/librados.hpp"
 // until everything is moved from rgw_common
