+++ src/rgw/rgw_rados.h
@@ -32,8 +32,8 @@
 #include "rgw_aio.h"
 #include "rgw_d3n_cacherequest.h"
 
+#include "services/rados/svc_bi_rados.h"
 #include "services/rados/svc_rados.h"
-#include "services/svc_bi_rados.h"
 #include "common/Throttle.h"
 #include "common/ceph_mutex.h"
 #include "rgw_cache.h"
