+++ src/rgw/services/rados/svc_bi_rados.cc
